Started refactoring utils
