+++ src/utils.h
@@ -9,7 +9,6 @@
 #include <iostream>
 #include <sstream>
 #include <cstdio>
-#include <algorithm>
 #include <cctype>
 #include <string>
 #include <cmath>
@@ -107,24 +106,6 @@ std::string to_str(const T value, const int n = 6)
     return out.str();
 }
 
-// Compares an extension with a list of extension strings
-// @return true if the extension matches one of those in the list
-bool checkExtension(const fs::path &extension, const std::initializer_list<std::string>& matches);
-
-time_t getModifiedTime(const std::string &filePath);
-off_t getSize(const std::string &filePath);
-bool pathsAreChildren(const fs::path &parentPath, const std::vector<std::string> &childPaths);
-bool pathIsChild(const fs::path &parentPath, const fs::path &p);
-int pathDepth(const fs::path &path);
-
-fs::path getExeFolderPath();
-fs::path getDataPath(const fs::path &p);
-fs::path getCwd();
-
-// Prints to the provided buffer a nice number of bytes (KB, MB, GB, etc)
-std::string bytesToHuman(off_t bytes);
-
-//
 std::string getPrompt(const std::string &prompt = "");
 
 // Cross-platform getpass
